backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0
